# fix(providers): GA optimized reset of custom dimensions and metrics