* **What kind of change does this PR introduce?**
It optimizes the set and reset of custom dimensions and metrics in google analytics


* **What is the current behavior?** (You can also link to an open issue here)
The code loops from 0 to 200 and sets all unused dimensions to undefined, same thing with metrics.  I assume this is to reset custom dimensions from a previous call.
This generates ~400 calls to ga('set', 'dimension-n', undefined) per each event that you want to track.
This takes between 2ms and 6 ms per event in a highspeed conection, if you want to track several events under a mobile connection it gets really slow.


* **What is the new behavior (if this is a feature change)?**
Instead of the looping from 0 to 200, the code caches setting custom dimensions and metrics and on next call unsets only the ones that will not be overriden with new values. This drastically speeds things up taking it from ~4ms to ~0.13 ms


* **Other information**: